[ES][v2] Move CoreDependencyStore and dbmodel from v1 to v2
