+++ .mockery.yaml
@@ -61,7 +61,7 @@ packages:
   github.com/jaegertracing/jaeger/internal/storage/v1/api/spanstore:
     config:
       all: true
-  github.com/jaegertracing/jaeger/internal/storage/v1/elasticsearch/dependencystore:
+  github.com/jaegertracing/jaeger/internal/storage/v2/elasticsearch/depstore:
     interfaces:
       CoreDependencyStore: {}
   github.com/jaegertracing/jaeger/internal/storage/v1/elasticsearch/spanstore:
