+++ internal/storage/v2/elasticsearch/depstore/dbmodel/converter.go
